Add reviews (#23)
changes in branch after review

+++ src/main/java/ru/yandex/practicum/filmorate/storages/ReviewStorage.java
@@ -5,21 +5,21 @@ import ru.yandex.practicum.filmorate.model.Review;
 import java.util.List;
 
 public interface ReviewStorage {
-    Review getReviewById(int reviewId);
+    Review getReviewById(Long reviewId);
 
     List<Review> getAllReviews();
 
-    List<Review> getFilmReviewsSortedByUsefulness(int filmId, int count);
+    List<Review> getFilmReviewsSortedByUsefulness(Long filmId, int count);
 
     Review addReview(Review review);
 
-    void addLikeOrDislikeToReview(int reviewId, int userId, boolean isLike);
+    void addLikeOrDislikeToReview(Long reviewId, Long userId, boolean isLike);
 
     Review updateReview(Review review);
 
-    void removeReview(int reviewId);
+    void removeReview(Long reviewId);
 
-    void removeLikeOrDislikeFromReview(int reviewId, int userId, boolean isLike);
+    void removeLikeOrDislikeFromReview(Long reviewId, Long userId, boolean isLike);
 
-    void assertReviewExists(int reviewId);
+    void assertReviewExists(Long reviewId);
 }
